reuse module org with manifest
